Update verify() to support other script types

+++ src/constants.jl
@@ -1,8 +1,8 @@
 # Script
 
-const SIGHASH_ALL = 1
-const SIGHASH_NONE = 2
-const SIGHASH_SINGLE = 3
+const SIGHASH_ALL = UInt32(1)
+const SIGHASH_NONE = UInt32(2)
+const SIGHASH_SINGLE = UInt32(3)
 
 const SCRIPT_TYPES = Dict([
     ("P2PKH", [0x6f, 0x00]),
